modify packets from uint8_t to uint16_t (follow ros)

+++ System/BRoCo/include/Protocol/ProtocolNUS24.h
@@ -131,8 +131,8 @@ RELIABLE_PACKET(competitionStatusPacket,
 )
 
 RELIABLE_PACKET(occupationStatusPacket,
-  uint8_t central_occupation;
-  uint8_t resupply_occupation;
+  uint16_t central_occupation;
+  uint16_t resupply_occupation;
 )
 
 RELIABLE_PACKET(winStatusPacket,
